fix 3 day button closes #40
