+++ src/components/Calendar/ThreeDayView/index.js
@@ -1,23 +1,53 @@
 import React from 'react'
 
 import dates from 'date-arithmetic'
-import events from '../events'
 import BigCalendar from 'react-big-calendar'
 import TimeGrid from 'react-big-calendar/lib/TimeGrid'
-import ExampleControlSlot from '../ExampleControlSlot'
+import moment from 'moment'
 
+import './index.scss';
 
-let CustomView = ({ localizer }) => (
-  <React.Fragment>
+// the css isn't applying for whatever reason
 
-    <BigCalendar
-      events={events}
-      localizer={localizer}
-      defaultView={BigCalendar.Views.WEEK}
-      defaultDate={new Date.today}
-      views={{ month: true, week: MyWeek }}
-    />
-  </React.Fragment>
-)
+class MyWeek extends React.Component {
+  render() {
+    let { date } = this.props
+    let range = MyWeek.range(date)
 
-export default CustomView
+    return <TimeGrid {...this.props} range={range} eventOffset={15} />
+  }
+}
+
+MyWeek.range = date => {
+  let start = date
+  let end = dates.add(start, 2, 'day')
+
+  let current = start
+  let range = []
+
+  while (dates.lte(current, end, 'day')) {
+    range.push(current)
+    current = dates.add(current, 1, 'day')
+  }
+
+  return range
+}
+
+MyWeek.navigate = (date, action) => {
+  switch (action) {
+    case BigCalendar.Navigate.PREVIOUS:
+      return dates.add(date, -3, 'day')
+
+    case BigCalendar.Navigate.NEXT:
+      return dates.add(date, 3, 'day')
+
+    default:
+      return date
+  }
+}
+
+MyWeek.title = date => {
+  return `${moment(date).format("MMM D")} - ${moment(dates.add(date, 3, 'day')).format("D")}`; 
+}
+
+export default MyWeek;
