+++ src/components/Calendar/Toolbar/index.scss
@@ -1,0 +1,7 @@
+.container .toolbar-container  {
+    width: 100%;
+}
+
+.btn .btn-default {
+    padding: 0;
+}
