+++ src/components/Calendar/index.js
@@ -1,10 +1,9 @@
 import React from 'react';
 // import ApiCalendar from 'react-google-calendar-api'; //https://www.npmjs.com/package/react-google-calendar-api
 import BigCalendar from 'react-big-calendar';
-
-import dates from 'date-arithmetic'
-import TimeGrid from 'react-big-calendar/lib/TimeGrid'
 import 'react-big-calendar/lib/css/react-big-calendar.css';
+import MyWeek from "./ThreeDayView";
+import CalendarToolbar from "./Toolbar"
 
 import './index.scss';
 
@@ -13,97 +12,6 @@ import moment from 'moment'
 // Setup the localizer by providing the moment (or globalize) Object
 // to the correct localizer.
 const localizer = BigCalendar.momentLocalizer(moment) // or globalizeLocalizer
-
-class MyWeek extends React.Component {
-  render() {
-    let { date } = this.props
-    let range = MyWeek.range(date)
-
-    return <TimeGrid {...this.props} range={range} eventOffset={15} />
-  }
-}
-
-MyWeek.range = date => {
-  let start = date
-  let end = dates.add(start, 2, 'day')
-
-  let current = start
-  let range = []
-
-  while (dates.lte(current, end, 'day')) {
-    range.push(current)
-    current = dates.add(current, 1, 'day')
-  }
-
-  return range
-}
-
-MyWeek.navigate = (date, action) => {
-  switch (action) {
-    case BigCalendar.Navigate.PREVIOUS:
-      return dates.add(date, -3, 'day')
-
-    case BigCalendar.Navigate.NEXT:
-      return dates.add(date, 3, 'day')
-
-    default:
-      return date
-  }
-}
-
-MyWeek.title = date => {
-  return `${moment(date).format("MMM D")} - ${moment(dates.add(date, 3, 'day')).format("D")}`; 
-}
- 
-// messing around with the toolbar
-// const CustomToolbar = (toolbar) => {
-//   let date = toolbar.date;
-
-//   let currentView = "week" ; // figure out how to get this view
-//   let jumps = currentView == "week" ? 3 : 7; 
-
-//   // const goToBack = () => {
-//   //   let newFirstDay = dates.add(date, -1*jumps, 'day')
-
-//   //   toolbar.date.setMonth(moment(newFirstDay).month);
-//   //   toolbar.onNavigate('prev');
-//   // };
-
-//   // const goToNext = () => {
-//   //   toolbar.date.setMonth(dates.add(date, jumps, 'day'));
-//   //   toolbar.onNavigate('next');
-//   // };
-
-//   // const goToCurrent = () => {
-//   //   const now = new Date();
-//   //   toolbar.date.setMonth(now.getMonth());
-//   //   toolbar.date.setYear(now.getFullYear());
-//   //   toolbar.onNavigate('current');
-//   // };
-
-//   const label = () => {
-//     const date = moment(toolbar.date);
-
-//     let end = dates.add(date, jumps, 'day');
-
-//     return (
-//       <span><b>{date.format('MM/DD/YY')}</b>-<b>{moment(end).format('MM/DD/YY')}</b></span>
-//     );
-//   };
-
-//   return (
-//     <div>
-//       <label>{label()}</label>
-
-//       <div >
-//         <button  onClick={BigCalendar.}>&#8249;</button>
-//         <button  onClick={goToCurrent}>today</button>
-//         <button  onClick={goToNext}>&#8250;</button>
-//       </div>
-//     </div >
-//   );
-// };
-
 
 export default class Calendar extends React.Component {
     constructor(props) {
@@ -118,8 +26,6 @@ export default class Calendar extends React.Component {
         })
       };
 
-
-
       //console.log("current user", currUser);
       // this.setState({currUser}, () => {console.log("done setting current user")});
       //console.log(this.props.user);
@@ -131,8 +37,6 @@ export default class Calendar extends React.Component {
 
     componentDidMount(){
       console.log("the calendar has mounted")
-      // add the user to the state
-      
 
       // get the current task list and turn them into calendar events
       let taskListRef = this.props.database.ref('taskList');
@@ -145,7 +49,6 @@ export default class Calendar extends React.Component {
     }
 
     tasksToEvents(){
-
       //console.log(this.state.tasks);
 
       let calEvents = Object.keys(this.state.tasks).map((key)=> {
@@ -166,7 +69,6 @@ export default class Calendar extends React.Component {
         if(this.state.tasks[key].groupID !== this.state.currUser.groupID){
           return {};
         }
-
 
         return {
           "start" : new Date(parseInt(this.state.tasks[key].taskDate)),
@@ -202,7 +104,6 @@ export default class Calendar extends React.Component {
           this.handleEventAdd(newEvent);
 
         }
-          
 
           // also need to add the one changed thing to the DB
 
@@ -221,8 +122,7 @@ export default class Calendar extends React.Component {
       //console.log("end", end);
       //console.log("isSelected", isSelected);
       //console.log("event assigned to ", event.assignedTo);
-      
-      // make the string Matt be user name.
+
       let backgroundColor = (this.state.currUser && event.assignedTo.includes(this.state.currUser.uid)) ? "#D66853" : "#96a6cc";
       let newStyle = {
           style : {backgroundColor}
@@ -231,9 +131,8 @@ export default class Calendar extends React.Component {
   }
   
 
-  
     handleEventAdd(event) {
-      console.log(event);
+      // console.log(event);
 
       let eventKey = this.props.database.ref().child('taskList').push().key;
       let submittedTask = {
@@ -261,18 +160,23 @@ export default class Calendar extends React.Component {
   }
 
     render() {
-      let components = { }
+      let components = {
+        toolbar: CalendarToolbar
+      };
 
       let formats = {
         dateFormat: 'dd D',
       
         dayFormat: (date, x, localizer) =>
-    localizer.format(date, 'dd D'),
+        localizer.format(date, 'dd D'),
       
+        timeGutterFormat: (date, x, localizer) =>
+        localizer.format(date, "h A"),
+      
+        dayRangeHeaderFormat: ({ start, end }, x, localizer) =>
+        `${localizer.format(start, "MMM D")} - ${localizer.format(end, "D")}` 
+      };
 
-      timeGutterFormat: (date, x, localizer) =>
-      localizer.format(date, "h A")
-      }
 
       return ( 
       <div className="cal-container">
@@ -282,9 +186,7 @@ export default class Calendar extends React.Component {
           formats={formats}          
           eventPropGetter={(e) => this.eventStyleGetter(e)}
           //toolbar={false}
-          // components={{
-          //   toolbar: CustomToolbar
-          // }}
+          components={components}
           events = {this.state.events}
           views={{week : true , MyWeek : MyWeek}}// make a custom view for three days to use for mobile
           showMultiDayTimes
@@ -292,7 +194,7 @@ export default class Calendar extends React.Component {
           defaultView={"week"}
           onSelectSlot={(slots) => {this.handleSelect(slots)}}
           popup = {true}
-          popupOffset={30}
+          //popupOffset={30}
           longPressThreshold={100}
           onSelectEvent= {(event, e) => this.handleSelectEvent(event, e)}
